--- src/app/programs/projects.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ProjectsService } from './projects.service';
-
-describe('ProjectsService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: ProjectsService = TestBed.get(ProjectsService);
-    expect(service).toBeTruthy();
-  });
-});
